[IE-229] Modularize window creation logic as a utility
